Rename record cursor class and factory to include AFTER as filter is applied after latest by
